--- src/components/BrownButton/BrownButton.js
@@ -1,30 +1,0 @@
-/* eslint-disable prettier/prettier */
-import React from 'react';
-import { View, Text, StyleSheet, Pressable} from 'react-native';
-
-const BrownButton = ({text, onPress}) => {
-
-  return (
-    <Pressable onPress={onPress} style={styles.container}>
-      <Text style={styles.text}>{text}</Text>
-    </Pressable>
-  );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#736468',
-    width: 150,
-    padding: 8.5,
-    marginVertical: 5,
-    alignItems: 'center',
-    borderRadius: 25,
-  },
-  text: {
-    fontWeight: 'bold',
-    color: 'white',
-    fontSize: 24,
-  },
-})
-
-export default BrownButton;
