+++ src/components/CustomButton/CustomButton.js
@@ -1,0 +1,46 @@
+/* eslint-disable prettier/prettier */
+import React from 'react';
+import { View, Text, StyleSheet, Pressable} from 'react-native';
+
+const CustomButton = ({text, onPress, type}) => {
+
+  return (
+    <Pressable onPress={onPress} style={[styles.container, styles[`container_${type}`]]}>
+      <Text style={styles.text, styles[`text_${type}`]}>{text}</Text>
+    </Pressable>
+  );
+};
+
+const styles = StyleSheet.create({
+  container_PRIMARY:{
+    backgroundColor: '#FFFFFF',
+    width: 150,
+    padding: 8.5,
+    marginVertical: 5,
+    alignItems: 'center',
+    borderRadius: 25,
+  },
+  container_SECONDARY:{
+    backgroundColor: '#736468',
+    width: 150,
+    padding: 8.5,
+    marginVertical: 5,
+    alignItems: 'center',
+    borderRadius: 25,
+  },
+  container_TERTIARY:{
+
+  },
+   text_PRIMARY:{
+    fontWeight: 'bold',
+    color: '#736468',
+    fontSize: 24,
+  },
+  text_SECONDARY: {
+    fontWeight: 'bold',
+    color: 'white',
+    fontSize: 24,
+  },
+})
+
+export default CustomButton;
