+++ src/components/CustomButton/index.js
@@ -1,0 +1,2 @@
+/* eslint-disable prettier/prettier */
+export { default } from './CustomButton';
