--- src/components/WhiteButton/index.js
@@ -1,2 +1,0 @@
-/* eslint-disable prettier/prettier */
-export { default } from './WhiteButton';
